+++ backend/processor/Processor.php
@@ -7,12 +7,19 @@ namespace mbolli\nfsen_ng\processor;
 /**
  * Provides access to netflow data using a model
  * compatible with nfdump commandline options.
+ *
+ * @phpstan-type ProcessorResult array{
+ *     command: string,
+ *     rawOutput: string,
+ *     decoded: array<mixed>,
+ *     stderr?: string,
+ * }
  */
 interface Processor {
     /**
      * Sets an option's value.
      *
-     * @param null|array|int|string $value
+     * @param null|array<mixed>|int|string $value
      */
     public function setOption(string $option, $value): void;
 
@@ -30,6 +37,8 @@ interface Processor {
     /**
      * Executes the processor command, tries to throw an
      * exception based on the return code.
+     *
+     * @return ProcessorResult
      *
      * @throws \Exception
      */